requirements.txt
torchmetrics
tensorboard
scipy
timm==0.4.12
yacs==0.1.8
wandb
einops
pytorch_warmup
ml_collections